Remove needless casting

+++ screen_brightness_ios/ios/Classes/SwiftScreenBrightnessIosPlugin.swift
@@ -86,7 +86,7 @@ public class SwiftScreenBrightnessIosPlugin: NSObject, FlutterPlugin, FlutterApp
     }
 
     public func handleCurrentBrightnessChanged(_ currentBrightness: CGFloat) {
-        currentBrightnessChangeStreamHandler.addCurrentBrightnessToEventSink(CGFloat(currentBrightness))
+        currentBrightnessChangeStreamHandler.addCurrentBrightnessToEventSink(currentBrightness)
     }
 
     @objc private func onSystemBrightnessChanged(notification: Notification) {
